+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- Add `alertmanagerconfig_matcher_strategy` to `mimir.alerts.kubernetes` component to change the matcher strategy. (@timonegk)
+- Add `alertmanagerconfig_matcher` block to `mimir.alerts.kubernetes` component to change the matcher strategy. (@timonegk)
 
 ## [1.12.0](https://github.com/jharvey10/test-repo/compare/v1.11.3...v1.12.0) (2025-12-01)